Move DatabaseItemImpl to separate file
